requirements.txt
# Caustic Water Removal - Requirements
# Install with: pip install -r requirements.txt
#
# For PyTorch with CUDA 12.1 (GPU recommended):
#   pip install torch==2.5.1+cu121 torchvision==0.20.1+cu121 torchaudio==2.5.1+cu121 \
#       --index-url https://download.pytorch.org/whl/cu121
#
# For CPU-only PyTorch:
#   pip install torch torchvision torchaudio

# Vision & Image processing
opencv-python>=4.13.0
Pillow>=11.3.0
scikit-image>=0.26.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0

# Scientific computing
numpy>=2.4.0
scipy>=1.17.0

# Video processing
moviepy>=2.2.0

# ML / Deep Learning (GPU version, see above for install command)
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1

# Utilities
tqdm>=4.67.0
matplotlib>=3.10.0
plyfile>=1.1.3
requests>=2.33.0
psutil>=7.2.0
python-dotenv>=1.2.0
